+++ CHANGELOG.md
@@ -7,5 +7,6 @@
 5. Updated Thoughts model to include Reactions
 6. Add User and Thought Routes
 7. Added Controllers - some not ready & so commented out
-8. Updated User Route & userController to add thought and update user
-9. Updated User Route & userController to add Friend
+8. Updated User Route & userController to add Thought and update user
+9. Updated User Route & userController to add a Friend
+10. Updated User Route & userController to remove a Friend